Extract "npm-dependency-version-checker" file
